LogoutInputData.java (should be done with the LogoutInteractor TODOs below)
TODO: save the current username in an instance variable and add a getter.
LogoutInteractor.java (tip: refer to ChangePasswordInteractor.java for similar code)

TODO: save the DAO and Presenter in the instance variables.
TODO: implement the logic of the Logout Use Case
LogoutOutputData.java

TODO: save the parameters in the instance variables.